+++ packages/next/src/index.ts
@@ -1,0 +1,4 @@
+export * from '@multitenant/core';
+export * from '@multitenant/config';
+export * from '@multitenant/react';
+export * from '@multitenant/next-app';